pipeline/requirements.txt
google-cloud-firestore>=2.14.0
google-cloud-secret-manager>=2.18.0
firebase-admin>=6.4.0
google-generativeai>=0.8.0
tenacity>=8.2.3
pydantic>=2.5.0

